+++ src/components/Info/info.scss
@@ -1,7 +1,8 @@
 .info {
   text-align: left;
   color: #000000;
-  h1 {}
+  h1 {
+  }
   .contents {
     font-size: 16px;
     overflow-y: scroll;
