+++ src/components/Layouts/InfoLayout/infoLayout.scss
@@ -7,6 +7,7 @@
   .contents_container {
     width: 100%;
     padding: 22px 32px;
+    background-color: #fff;
   }
   .menu_container {
     display: flex;
