+++ src/components/Main/Main.scss
@@ -1,0 +1,19 @@
+.mainpage {
+  display: flex;
+  flex-flow: row nowrap;
+  margin: 0 0;
+  padding: 80px 212px 140px;
+  text-align: center;
+
+  .background {
+    width: 100%;
+    position: fixed;
+    left: 0;
+    bottom: 0;
+    right: 0;
+    z-index: -1;
+  }
+  .main_container {
+    width: 980px;
+  }
+}
