+++ src/components/Main/MainSide/MainSide.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import './mainSide.scss'
+
+const MainSide = () => {
+  return <div className="main_side">Side Menu</div>
+}
+
+export default MainSide
