+++ src/components/Main/MainSide/index.js
@@ -1,0 +1,1 @@
+export { default } from './MainSide'
