+++ src/components/Main/MainSide/mainSide.scss
@@ -1,0 +1,6 @@
+.main_side {
+  background-color: #eeeeee;
+  width: 480px;
+  margin-left: 32px;
+  border: 1px solid #818181;
+}
