--- src/pages/mainPage.scss
@@ -1,19 +1,0 @@
-.mainpage {
-  display: flex;
-  flex-flow: row nowrap;
-  margin: 0 0;
-  padding: 80px 212px 140px;
-  text-align: center;
-
-  .background {
-    width: 100%;
-  }
-  .main_container {
-    width: 980px;
-  }
-  .side_container {
-    width: 480px;
-    margin-left: 32px;
-    border: 1px solid #818181;
-  }
-}
